requirements.txt
Flask
gunicorn
python-dotenv
SQLAlchemy
pandas
openpyxl
psycopg2-binary
psycopg[binary]